refactor: custom spotify hook
